requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
duckdb==1.4.3
idna==3.11
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.2
